feat: reformat code into multiple files
